+++ src/class.rs
@@ -1,6 +1,6 @@
 use crate::StockWeapon;
 use serde::{Deserialize, Serialize};
-use strum_macros::{Display, EnumString, EnumIter, EnumCount};
+use strum::{Display, EnumString, EnumIter, EnumCount};
 
 /// Class.
 #[derive(