+++ src/error.rs
@@ -1,6 +1,8 @@
 use crate::Spell;
 use std::fmt;
 
+pub use num_enum::TryFromPrimitiveError;
+
 /// An error when attempting to convert a spell into a sub-set of spells (footprints or paint).
 #[derive(Debug)]
 pub struct TryFromSpellError {