+++ src/item_level.rs
@@ -1,7 +1,7 @@
 use crate::KillEaterScoreType;
 use std::fmt;
 use serde::{Deserialize, Serialize};
-use strum_macros::{Display, EnumString, EnumIter, EnumCount};
+use strum::{Display, EnumString, EnumIter, EnumCount};
 
 /// Item level.
 #[derive(