+++ src/strange_part.rs
@@ -1,5 +1,6 @@
 use crate::{KillEaterScoreType, Attributes, ItemDefindex, EffectType, DescriptionFormat};
-use strum_macros::{Display, EnumString, EnumIter, EnumCount};
+use crate::error::TryFromPrimitiveError;
+use strum::{Display, EnumString, EnumIter, EnumCount};
 use num_enum::{TryFromPrimitive, IntoPrimitive};
 use serde_repr::{Serialize_repr, Deserialize_repr};
 
@@ -438,7 +439,7 @@ impl ItemDefindex for StrangePart {
 }
 
 impl TryFrom<KillEaterScoreType> for StrangePart {
-    type Error = num_enum::TryFromPrimitiveError<StrangePart>;
+    type Error = TryFromPrimitiveError<StrangePart>;
     
     fn try_from(part: KillEaterScoreType) -> Result<Self, Self::Error> {
         StrangePart::try_from(part as u32)
@@ -446,7 +447,7 @@ impl TryFrom<KillEaterScoreType> for StrangePart {
 }
 
 impl TryFrom<&KillEaterScoreType> for StrangePart {
-    type Error = num_enum::TryFromPrimitiveError<StrangePart>;
+    type Error = TryFromPrimitiveError<StrangePart>;
     
     fn try_from(part: &KillEaterScoreType) -> Result<Self, Self::Error> {
         StrangePart::try_from(*part as u32)